move files into the V1 spec and set API version to 1
