+++ LBHTenancyAPI/Controllers/V1/ServiceController.cs
@@ -1,11 +1,15 @@
 using System.Threading.Tasks;
-using LBHTenancyAPI.Controllers.V1;
 using LBHTenancyAPI.Extensions.Controller;
+using LBHTenancyAPI.Infrastructure.V1.API;
 using LBHTenancyAPI.UseCases.Service;
 using Microsoft.AspNetCore.Mvc;
 
-namespace LBHTenancyAPI.Controllers
+namespace LBHTenancyAPI.Controllers.V1
 {
+    [ApiVersion("1")]
+    [Produces("application/json")]
+    [ProducesResponseType(typeof(APIResponse<object>), 400)]
+    [ProducesResponseType(typeof(APIResponse<object>), 500)]
     public class ServiceController : BaseController
     {
         private readonly IGetServiceDetailsUseCase _serviceDetailsUseCase;
@@ -16,6 +20,7 @@ namespace LBHTenancyAPI.Controllers
         }
 
         [HttpGet]
+        [ProducesResponseType(typeof(APIResponse<GetServiceDetailsResponse>), 200)]
         [Route("/")]
         public async Task<IActionResult> Get()
         {
